WIP: Removes subcommand from 'mc admin info <subcmd> <alias>'
